Initial implementation of task name in navbar
Added a text field in the same div as navbar-right-actions.
Updating of the text is done in taskNameDisplay.js and is called in two locations in min, 1. when switching a task, and the nametask bang. NB: Renaming the task in the task overlay will automatically call the switchToTask, and therefore update this text field.